refactor(modularization): Phase 3 Step 4b — extract 11 KB schemas + re-export shim + shared common (#16)
Per @符炫炜 msg=1505044c / msg=eb271056 / msg=618c03fc Option X Step 4b
canonical α. Moves the 11 Phase 3 schemas out of
`aperag.schema.view_models` into
`aperag.domains.knowledge_base.schemas` so the KB domain owns its
public contract shape. View_models re-exports them for back-compat
pending the Phase 6 cleanup sweep.

Sub-edits bundled in this atomic commit:

1. New `aperag/schema/common.py` hosts the shared primitives that
   the KB schemas transitively depend on (ModelSpec,
   KnowledgeGraphConfig, IndexPrompts, CollectionConfig, PageResult,
   PaginatedResponse, Chunk, VisionChunk). These types are also used
   by non-KB domains (source / bots / retrieval / views), so keeping
   them in `aperag.schema.view_models` would trip Phase 3 G1
   (legacy-aggregate ban) when the KB domain imports them. The new
   module is intentionally outside the G1 ban list.

2. New `aperag/domains/knowledge_base/{__init__.py,schemas.py}` with
   the 11 canonical KB schemas (Collection, CollectionView,
   CollectionViewList, CollectionCreate, CollectionUpdate, Document,
   DocumentList, DocumentPreview, RebuildIndexesRequest,
   RebuildIndexesResponse, CollectionSummaryTriggerResponse). The
   schemas depend only on `aperag.schema.common`; G1 boundary test
   stays green.

3. `aperag/schema/view_models.py`:
   * deletes the 11 KB class definitions and the 8 shared-common
     definitions that moved to common.py;
   * imports the 8 shared names back from common.py as re-exports
     (back-compat for callers like `aperag.source.base`,
     `aperag.views.utils`, `aperag.service.chat_service` etc.);
   * end-of-file `try/except ImportError` imports the 11 KB schemas
     from `aperag.domains.knowledge_base.schemas` so
     `from aperag.schema.view_models import Collection` and
     Pydantic forward-ref resolution (e.g. `Agent.collections:
     list[Collection]`) continue to see the canonical class objects.
   * a symmetric `_bind_view_models_reexports` hook lives at the end
     of `knowledge_base/schemas.py` to handle the KB-first module
     load order (view_models.py end import hits the circular window
     and its try/except swallows the ImportError; the KB hook binds
     the names back via `sys.modules["aperag.schema.view_models"]`
     so the AST-level G1 scan stays clean — no literal
     `import aperag.schema.view_models` inside the KB module).

4. External caller rewrites — 4 files imported moved schemas via
   `from aperag.schema.view_models import`; rewrite to
   `from aperag.domains.knowledge_base.schemas import`:
     - aperag/mcp/server.py (CollectionViewList)
     - aperag/service/document_service.py (DocumentList, DocumentPreview)
     - aperag/service/collection_service.py (Collection)
     - aperag/service/chat_collection_service.py (CollectionCreate)
   `view_models.Collection` / `view_models.Document` style
   attribute access (used in ~50 other call sites like
   `aperag/service/marketplace_collection_service.py`) still works
   through the re-export shim — those are deferred to Phase 6.

### Verification

- `uv run pytest tests/unit_test/`: 669 passed, 29 skipped
  (G1 / G11 15-symbol audit / G13 Phase-3-completion / G14 Protocol
  direction all green)
- `make lint`: all checks passed
- `make openapi-check`: empty diff (no OpenAPI shape change — same
  class objects are referenced via re-export chain)
- Import-identity smoke: `vm.Collection is kb.Collection` = True for
  both load orders (view_models-first and KB-first); every entry in
  `__all__` passes the identity check.
- G12 `alembic check` is CI-side (no local Docker); pure Python
  refactor with no schema touches, empty diff expected.

### Canonical adherence

- Option X 8-step Step 4 sub-slice 4b (no views/services/router move)
- Phased re-export remains 15/15 final (no change from Step 4a)
- G11 `test_aperag_db_models_reexports_full_phase3_set` continues
  to pass
- `view_models.py` shrinks 1846 → 1478 lines (~20% reduction in
  the god module; further cleanup in Phase 6)

### Non-goals deferred

- `view_models.Chunk`, `VisionChunk`, `PageResult`,
  `PaginatedResponse`, `ModelSpec`, `KnowledgeGraphConfig`,
  `IndexPrompts`, `CollectionConfig` remain re-exported through
  view_models for back-compat; Phase 6 will rewrite the ~60 callers
  and drop the re-export shim.
- Service / view / app.py router moves are Step 5 (Bryce returning).
