+++ aperag/domains/knowledge_base/__init__.py
@@ -1,0 +1,21 @@
+# Copyright 2025 ApeCloud, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Canonical `knowledge_base` domain package.
+
+Phase 3 Step 4b (msg=1505044c) introduces
+:mod:`aperag.domains.knowledge_base.schemas` as the owner of the
+collection- and document-facing Pydantic view models. See
+``notes/modularization-v2.md`` for the Option X 8-step Phase 3 plan.
+"""
